PUT command update
Has been removed the list and improved the performance of the code. Also itemContainer is an Entity class due to control the behaviour of the command. For example, if it puts the name of an NPC in the container slot of the command and finds the NPC in the room.